+++ test/access_token.js
@@ -2,70 +2,51 @@
 
 const qs = require('querystring');
 const nock = require('nock');
+const path = require('path');
+const expect = require('chai').expect;
 const startOfYesterday = require('date-fns/start_of_yesterday');
 const oauth2Module = require('./../index.js');
+const isValid = require('date-fns/is_valid');
+const isEqual = require('date-fns/is_equal');
 
-const oauth2 = oauth2Module({
-  clientID: 'client-id',
-  clientSecret: 'client-secret',
-  site: 'https://example.org',
-});
+const oauth2 = oauth2Module
+  .create(require('./fixtures/oauth-options'));
 
-let request;
-let result;
-let resultPromise;
-let token;
-let tokenPromise;
-let error;
-let errorPromise;
-const tokenConfig = {
+const tokenParams = {
   code: 'code',
   redirect_uri: 'http://callback.com',
 };
-const refreshConfig = {
-  grant_type: 'refresh_token',
-  refresh_token: 'ec1a59d298',
-  client_id: 'client-id',
-  client_secret: 'client-secret',
-};
-const refreshWithAdditionalParamsConfig = {
-  scope: 'TESTING_EXAMPLE_SCOPES',
-  grant_type: 'refresh_token',
-  refresh_token: 'ec1a59d298',
-  client_id: 'client-id',
-  client_secret: 'client-secret',
-};
-const revokeConfig = {
-  token: 'ec1a59d298',
-  token_type_hint: 'refresh_token',
-  client_id: 'client-id',
-  client_secret: 'client-secret',
-};
-const oauthConfig = {
-  code: 'code',
-  redirect_uri: 'http://callback.com',
-  grant_type: 'authorization_code',
-  client_id: 'client-id',
-  client_secret: 'client-secret',
-};
+
+const refreshConfig = require('./fixtures/refresh-token.json');
+const refreshWithAdditionalParamsConfig = require('./fixtures/refresh-token-with-params.json');
+const oauthConfig = require('./fixtures/oauth-options-code.json');
+const revokeConfig = require('./fixtures/revoke-token-params.json');
 
 describe('oauth2.accessToken', function () {
+  let request;
+  let result;
+  let resultPromise;
+  let token;
+  let tokenPromise;
+  let error;
+  let errorPromise;
+
   beforeEach(function () {
     request = nock('https://example.org:443')
       .post('/oauth/token', qs.stringify(oauthConfig))
       .times(2)
-      .replyWithFile(200, __dirname + '/fixtures/access_token.json');
+      .replyWithFile(200, path.join(__dirname, '/fixtures/access_token.json'));
   });
 
   beforeEach(function (done) {
-    oauth2.authCode.getToken(tokenConfig, function (e, r) {
+    oauth2.authorizationCode.getToken(tokenParams, function (e, r) {
       error = e; result = r; done();
     });
   });
 
   beforeEach(function () {
-    return oauth2.authCode
-      .getToken(tokenConfig)
+    return oauth2.authorizationCode
+      .getToken(tokenParams)
       .then(function (r) { resultPromise = r; })
       .catch(function (e) { errorPromise = e; });
   });
@@ -76,19 +57,39 @@ describe('oauth2.accessToken', function () {
   });
 
   describe('#create', function () {
-    it('creates an access token as result of callback api', function () {
-      token.should.have.property('token');
+    it('creates an access token wrapper object', function () {
+      expect(token).to.have.property('token');
+      expect(tokenPromise).to.have.property('token');
+    });
+  });
+
+  describe('#create with expires_at', function () {
+    it('uses the set expires_at property', function () {
+      token.token.expires_at = startOfYesterday();
+      const expiredToken = oauth2.accessToken.create(token.token);
+      expect(isValid(expiredToken.token.expires_at)).to.be.equal(true);
+      expect(isEqual(expiredToken.token.expires_at, token.token.expires_at))
+        .to.be.equal(true);
     });
 
-    it('created an access token as result of promise api', function () {
-      tokenPromise.should.have.property('token');
+    it('parses a set expires_at property', function () {
+      const yesterday = startOfYesterday();
+      token.token.expires_at = yesterday.toString();
+      const expiredToken = oauth2.accessToken.create(token.token);
+      expect(isValid(expiredToken.token.expires_at)).to.be.equal(true);
+      expect(isEqual(expiredToken.token.expires_at, token.token.expires_at))
+        .to.be.equal(true);
+    });
+
+    it('create its own date by default', function () {
+      expect(isValid(token.token.expires_at)).to.be.equal(true);
     });
   });
 
   describe('when not expired', function () {
     it('returns false', function () {
-      token.expired().should.be.false;
-      tokenPromise.expired().should.be.false;
+      expect(token.expired()).to.be.equal(false);
+      expect(tokenPromise.expired()).to.be.equal(false);
     });
   });
 
@@ -99,8 +100,8 @@ describe('oauth2.accessToken', function () {
     });
 
     it('returns false', function () {
-      token.expired().should.be.true;
-      tokenPromise.expired().should.be.true;
+      expect(token.expired()).to.be.equal(true);
+      expect(tokenPromise.expired()).to.be.equal(true);
     });
   });
 
@@ -109,7 +110,7 @@ describe('oauth2.accessToken', function () {
       request = nock('https://example.org:443')
         .post('/oauth/token', qs.stringify(refreshConfig))
         .times(2)
-        .replyWithFile(200, __dirname + '/fixtures/access_token.json');
+        .replyWithFile(200, path.join(__dirname, '/fixtures/access_token.json'));
     });
 
     beforeEach(function (done) {
@@ -129,17 +130,14 @@ describe('oauth2.accessToken', function () {
     });
 
     it('makes the HTTP request', function () {
-      request.isDone().should.be.true;
+      expect(request.isDone()).to.be.equal(true);
     });
 
-    it('returns a new oauth2.accessToken as result of callback api', function () {
-      result.should.not.be.equal(global);
-      result.token.should.have.property('access_token');
-    });
-
-    it('returns a new oauth2.accessToken as result of promise api', function () {
-      result.should.not.be.equal(global);
-      resultPromise.token.should.have.property('access_token');
+    it('returns a new oauth2.accessToken as a result of the token refresh', function () {
+      expect(result).to.not.be.equal(global);
+      expect(result.token).to.have.property('access_token');
+      expect(resultPromise).to.not.be.equal(global);
+      expect(resultPromise.token).to.have.property('access_token');
     });
   });
 
@@ -148,7 +146,7 @@ describe('oauth2.accessToken', function () {
       request = nock('https://example.org:443')
         .post('/oauth/token', qs.stringify(refreshWithAdditionalParamsConfig))
         .times(2)
-        .replyWithFile(200, __dirname + '/fixtures/access_token.json');
+        .replyWithFile(200, path.join(__dirname, '/fixtures/access_token.json'));
     });
 
     beforeEach(function (done) {
@@ -168,15 +166,16 @@ describe('oauth2.accessToken', function () {
     });
 
     it('makes the HTTP request', function () {
-      request.isDone().should.be.true;
+      expect(request.isDone()).to.be.equal(true);
     });
 
     it('returns a new oauth2.accessToken as result of callback api', function () {
-      result.token.should.have.property('access_token');
+      expect(result.token).to.have.property('access_token');
     });
 
-    it('returns a new oauth2.accessToken as result of promise api', function () {
-      resultPromise.token.should.have.property('access_token');
+    it('returns a new oauth2.accessToken as a result of the token refresh', function () {
+      expect(result.token).to.have.property('access_token');
+      expect(resultPromise.token).to.have.property('access_token');
     });
   });
 
@@ -199,13 +198,13 @@ describe('oauth2.accessToken', function () {
       resultPromise = null;
       errorPromise = null;
 
-      return tokenPromise.revoke()
+      return tokenPromise.revoke('refresh_token')
         .then(function (r) { resultPromise = r; })
         .catch(function (e) { errorPromise = e; });
     });
 
     it('make HTTP call', function () {
-      request.isDone().should.be.true;
+      expect(request.isDone()).to.be.equal(true);
     });
   });
 });