+++ test/password.js
@@ -1,25 +1,19 @@
 'use strict';
 
-require('should');
-const oauth2Module = require('./../index.js');
+const path = require('path');
 const qs = require('querystring');
 const nock = require('nock');
+const expect = require('chai').expect;
+const oauth2Module = require('./../index.js');
 
-const oauth2 = oauth2Module({
-  clientID: 'client-id',
-  clientSecret: 'client-secret',
-  site: 'https://example.org',
-});
+const oauth2 = oauth2Module
+  .create(require('./fixtures/oauth-options'));
 
-let request;
-let result;
-let resultPromise;
-let error;
-let errorPromise;
 const tokenParams = {
   username: 'alice',
   password: 'secret',
 };
+
 const oauthParams = {
   username: 'alice',
   password: 'secret',
@@ -30,35 +24,38 @@ const oauthParams = {
 
 describe('oauth2.password', function () {
   describe('#getToken', function () {
+    let request;
+    let result;
+    let resultPromise;
+    let error;
+    let errorPromise;
+
     beforeEach(function () {
       request = nock('https://example.org:443')
         .post('/oauth/token', qs.stringify(oauthParams))
         .times(2)
-        .replyWithFile(200, __dirname + '/fixtures/access_token.json');
+        .replyWithFile(200, path.join(__dirname, '/fixtures/access_token.json'));
     });
 
     beforeEach(function (done) {
-      oauth2.password.getToken(tokenParams, function (e, r) {
+      oauth2.ownerPassword.getToken(tokenParams, function (e, r) {
         error = e; result = r; done();
       });
     });
 
     beforeEach(function () {
-      return oauth2.password.getToken(tokenParams)
+      return oauth2.ownerPassword.getToken(tokenParams)
         .then(function (r) { resultPromise = r; })
         .catch(function (e) { errorPromise = e; });
     });
 
     it('makes the HTTP request', function () {
-      request.isDone().should.be.true;
+      expect(request.isDone()).to.be.equal(true);
     });
 
-    it('returns an access token as result of callback api', function () {
-      result.should.have.property('access_token');
-    });
-
-    it('returns an access token as result of promises api', function () {
-      resultPromise.should.have.property('access_token');
+    it('returns an access token as result of the token request', function () {
+      expect(result).to.have.property('access_token');
+      expect(resultPromise).to.be.property('access_token');
     });
   });
 });